+++ src/ClientApi/Rest/Exception/UnprocessableEntityException.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: jjuszkiewicz
+ * Date: 14.07.15
+ * Time: 08:46
+ */
+
+namespace Nokaut\ApiKit\ClientApi\Rest\Exception;
+
+
+class UnprocessableEntityException extends \Exception {
+
+}